deepseek_pricing/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.25.0
python-dotenv>=0.19.0
markdown>=3.3.0
argparse==1.4.0 